+++ packages/block-library/src/playlist/context.js
@@ -6,5 +6,4 @@ import { createContext } from '@wordpress/element';
 export const PlaylistContext = createContext( {
 	currentTrackClientId: null,
 	setCurrentTrackClientId: () => {},
-	addTracks: undefined,
 } );